Fix for pylint
