+++ tests/workspace_test.py
@@ -36,7 +36,7 @@ def test_workspace_from_kwargs(tmp_path: Path):
         "distance_unit": "feet",
     }
 
-    workspace = Workspace(**attr).save_as(h5file_tmp)
+    workspace = Workspace(**attr).save_as(h5file_tmp)  # type: ignore
 
     # Test re-opening in read-only - stays in r+"
     with pytest.warns(UserWarning) as warning:
@@ -113,7 +113,7 @@ def test_read_bytes(tmp_path):
     with pytest.warns(UserWarning, match=""):
         workspace = Workspace(tmp_path / r"test_warning.geoh5")
 
-    assert workspace.h5file.is_file()
+    assert workspace.h5file.is_file()  # type: ignore
 
     with Workspace.create(tmp_path / r"test.geoh5") as workspace:
         workspace.create_entity(Points)
@@ -133,6 +133,9 @@ def test_read_bytes(tmp_path):
 def test_reopening_mode(tmp_path):
     with Workspace.create(tmp_path / r"test.geoh5") as workspace:
         pass
+
+    with fetch_active_workspace(workspace, mode="r") as re_open:
+        assert re_open.geoh5.mode == "r"
 
     with workspace.open(mode="r") as workspace:
         with fetch_active_workspace(workspace, mode="r") as re_open:
